# openapi/local-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CSGHub Lite Local API",
    "version": "0.1.0",
    "description": "Structured documentation for the local `/api/*` and `/v1/*` endpoints served by `csghub-lite`.\n\nMaintenance note: whenever you add, remove, or rename an `/api/*` or `/v1/*` route in `internal/server/routes.go`, update this file and run the OpenAPI sync test before merging."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current csghub-lite server"
    },
    {
      "url": "http://localhost:11435",
      "description": "Default local server"
    }
  ],
...
  "paths": {
...
    "/api/tags": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List available models",
        "description": "Returns local, CSGHub cloud, and selected third-party provider models. Each model preserves its configured provider display name while `source` carries its stable routing identity. Use `provider` to filter by a stable provider ID or configured provider name. Provider filters return selected models configured through `/api/tags/manage`; when no `csghub` cloud selection has been saved, all cloud catalog models are returned.",
        "operationId": "getTags",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "description": "Force a refresh of the cloud model catalog before returning results.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "description": "Filter models case-insensitively by provider alias. Local models accept `local`; cloud models accept `csghub` or the configured cloud provider name; third-party models accept their stable provider ID or configured provider name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Optional csghub-lite category key from `/api/pipeline-tags`. When provided, only models whose pipeline tag belongs to the category are returned.",
            "schema": {
              "type": "string",
              "enum": [
                "language_model",
                "image_generation",
                "video_generation",
                "text_to_speech",
                "speech_recognition"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/tags/manage": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List provider catalog models",
        "description": "Returns all currently available models from the built-in `csghub` cloud provider or a configured third-party provider. This catalog is used to choose which provider models appear in `/api/tags?provider=...`.",
        "operationId": "listProviderTagCatalog",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Provider ID or configured provider name. Use `csghub` for the built-in cloud provider.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Optional csghub-lite category key from `/api/pipeline-tags`. When omitted, all provider catalog models are returned.",
            "schema": {
              "type": "string",
              "enum": [
                "language_model",
                "image_generation",
                "video_generation",
                "text_to_speech",
                "speech_recognition"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider catalog models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Select a provider model",
        "operationId": "addProviderTagModel",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Provider ID or configured provider name. Use `csghub` for the built-in cloud provider.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTagModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model was already selected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            }
          },
          "201": {
            "description": "Model selected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Replace selected provider models",
        "operationId": "replaceProviderTagModels",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Provider ID or configured provider name. Use `csghub` for the built-in cloud provider.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Optional csghub-lite category key from `/api/pipeline-tags`. When supplied, the response is filtered to selected models in that category.",
            "schema": {
              "type": "string",
              "enum": [
                "language_model",
                "image_generation",
                "video_generation",
                "text_to_speech",
                "speech_recognition"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTagModelsReplaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Selected provider models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Models"
        ],
        "summary": "Update selected provider model metadata",
        "description": "Updates metadata for a selected provider model and can change the client-visible model ID. The built-in `csghub` cloud provider also supports this endpoint; patching a default cloud catalog model creates a selection automatically. The original upstream model ID is preserved internally for inference calls. Use an empty string to clear an overridden display name or description.",
        "operationId": "updateProviderTagModel",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Provider ID or configured provider name. Use `csghub` for the built-in cloud provider.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Current client-visible provider model ID to update. Required when changing the model ID; the request body `model` value is the new client-visible ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTagModelUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated provider model metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "Models"
        ],
        "summary": "Unselect a provider model",
        "operationId": "deleteProviderTagModel",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "description": "Provider ID or configured provider name. Use `csghub` for the built-in cloud provider.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": true,
            "description": "Provider model ID to remove from the selected list.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model unselected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/models/search": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Search local downloaded models",
        "operationId": "searchLocalModels",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive search text matched against model metadata.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Filter by local model format, such as `gguf`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pipeline_tag",
            "in": "query",
            "description": "Filter by detected pipeline type, such as `text-generation` or `image-text-to-text`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return. Defaults to `20`, capped at `100`.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Zero-based offset into the filtered result set.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching local models.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalModelSearchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/models/{model}/manifest": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get a local model download manifest by public model ID",
        "description": "Resolves a local model by its public inference ID, such as a short model name shown in the local model library, and returns its download manifest.",
        "operationId": "getModelManifestByPublicID",
        "parameters": [
          {
            "name": "model",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Local model metadata with downloadable files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelManifestResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/models/{namespace}/{name}/manifest": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get a local model download manifest",
        "operationId": "getModelManifest",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Local model metadata with downloadable files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelManifestResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/models/upload": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Upload a local model",
        "description": "Imports a local model from multipart form data. Use `mode=archive` for a single `.zip`, `.tar`, `.tar.gz`, or `.tgz` file; use `mode=files` or `mode=directory` with one or more files and matching `paths` values.",
        "operationId": "uploadModel",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "files"
                ],
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Optional target model ID in `namespace/name` form. Defaults to `local/<filename>` when omitted."
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "archive",
                      "directory",
                      "files"
                    ],
                    "default": "files"
                  },
                  "overwrite": {
                    "type": "boolean",
                    "default": false
                  },
                  "paths": {
                    "type": "array",
                    "description": "Relative paths corresponding to uploaded files. Folder uploads should provide browser-relative paths.",
                    "items": {
                      "type": "string"
                    }
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The imported local model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelUploadResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/models/upload/start": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Start a local model upload session",
        "description": "Creates a resumable-style upload session used by the web UI for large model files. Files are uploaded one at a time with raw PUT requests, then imported on complete.",
        "operationId": "startModelUpload",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelUploadStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload session created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelUploadStartResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/upload/{uploadID}/file": {
      "put": {
        "tags": [
          "Models"
        ],
        "summary": "Upload a file to a model upload session",
        "description": "Uploads a model file as raw bytes. For large files, send sequential chunks with `Content-Range: bytes start-end/total`; the server validates and returns the next accepted offset. Omitting `Content-Range` preserves whole-file PUT compatibility.",
        "operationId": "uploadModelSessionFile",
        "parameters": [
          {
            "name": "uploadID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Content-Range",
            "in": "header",
            "required": false,
            "description": "Byte range for this chunk, for example `bytes 0-33554431/1073741824`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File or chunk accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelUploadChunkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "The supplied chunk offset does not match the stored file offset."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/models/upload/{uploadID}/complete": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Complete a model upload session",
        "operationId": "completeModelUpload",
        "parameters": [
          {
            "name": "uploadID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The imported local model.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelUploadResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/models/upload/{uploadID}": {
      "delete": {
        "tags": [
          "Models"
        ],
        "summary": "Cancel a model upload session",
        "operationId": "cancelModelUpload",
        "parameters": [
          {
            "name": "uploadID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Upload session cancelled."
          }
        }
      }
    },
...
    "/api/datasets": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "List local datasets",
        "operationId": "getDatasets",
        "responses": {
          "200": {
            "description": "Available local datasets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetTagsResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "The request body, query parameters, or path parameters were invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested model, dataset, app, or shell session was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An internal server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "TagsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "models"
        ],
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            }
          }
        }
      },
...
    }
  }
}